+++ argon2pure.py
@@ -11,6 +11,7 @@ from six import BytesIO
 
 import struct
 import binascii
+import multiprocessing
 
 __all__ = [
     'argon2',
@@ -30,7 +31,7 @@ class Argon2ParameterError(Argon2Error):
 
 def argon2(password, salt, time_cost, memory_cost, parallelism,
                 tag_length=32, secret=b'', associated_data=b'',
-                type_code=ARGON2I):
+                type_code=ARGON2I, threads=None):
     """ Compute the Argon2 hash for *password*.
 
     :param bytes password: Password to hash
@@ -47,17 +48,31 @@ def argon2(password, salt, time_cost, memory_cost, parallelism,
     :param bytes secret: Optional secret to differentiate hash
     :param bytes associated_data: Optional associated data
     :param int type: variant of argon2 to use.  Either ARGON2I or ARGON2D
+    :param int threads: number of threads to use to compute the hash.
 
     :rtype: bytes """
-    # Compute the pre-hasing digest
-    if parallelism < 0:
+    if threads is None:
+        threads = parallelism
+    if parallelism <= 0:
         raise Argon2ParameterError("parallelism must be strictly positive")
-    if time_cost < 0:
+    if threads <= 0:
+        raise Argon2ParameterError("threads must be strictly positive")
+    if time_cost <= 0:
         raise Argon2ParameterError("time_cost must be strictly positive")
     if memory_cost < 8 * parallelism:
         raise Argon2ParameterError("memory_cost can't be less than 8"
                                     " times the number of lanes")
+    if type_code not in (0, 1):
+        raise Argon2ParameterError("type_code %s not supported" % type_code)
 
+    threads = min(parallelism, threads)
+
+    if threads == 1:
+        worker_pool = None
+    else:
+        worker_pool = multiprocessing.Pool(processes=threads)
+
+    # Compute the pre-hasing digest
     h = Blake2b()
     h.update(struct.pack("<iiiiii", parallelism,
                                     tag_length,
@@ -79,9 +94,6 @@ def argon2(password, salt, time_cost, memory_cost, parallelism,
     q = m_prime // parallelism  # lane_length
     segment_length = q // 4
 
-    if type_code not in (0, 1):
-        raise Argon2ParameterError("type_code %s not supported" % type_code)
-
     # Allocate the matrix.
     B = [[None for j in range(q)] for i in range(parallelism)]
 
@@ -95,9 +107,21 @@ def argon2(password, salt, time_cost, memory_cost, parallelism,
     # a slice can be computed in parallel.
     for t in range(time_cost):
         for segment in range(4):
+            if not worker_pool:
+                for i in range(parallelism):
+                    _fill_segment(B, t, segment, i, type_code, segment_length,
+                                    H0, q, parallelism, m_prime, time_cost)
+                continue
+
+            handles = [None]*parallelism
             for i in range(parallelism):
-                _fill_segment(B, t, segment, i, type_code, segment_length, H0,
-                                    q, parallelism, m_prime, time_cost)
+                handles[i] = worker_pool.apply_async(_fill_segment,
+                                (B, t, segment, i, type_code, segment_length,
+                                    H0, q, parallelism, m_prime, time_cost))
+            for i in range(parallelism):
+                new_blocks = handles[i].get()
+                for index in range(segment_length):
+                    B[i][segment * segment_length + index] = new_blocks[index]
 
     B_final = b'\0' * 1024
 
@@ -171,6 +195,9 @@ def _fill_segment(B, t, segment, i, type_code, segment_length, H0,
         # Mix the previous and reference block to create
         # the next block.
         B[i][j] = _compress(B[i][(j-1)%q], B[i_prime][j_prime])
+
+    # If we are run in a separate thread, then B is a copy.  Return changes.
+    return B[i][segment*segment_length:(segment+1)*segment_length]
 
 
 # xor1024: XOR two 1024 byte blocks with eachother.